Check for various extras keys: default, message, and optional title
The `message` key is now used for the `ticker` text when a notification arrives.

A `default` key that's sent by Amazon SNS for plain text pushes will be used as the `message` if it's present.

The `title` key is now optional. If not sent, the app name will be used as the `title` text.